segregation of bad complexes + cleanup

+++ Scripts/qcgen.py
@@ -1,4 +1,5 @@
-# Written by Tim Ioannidis for HJK Group
+# Written by Tim Ioannidis for HJK
+#         and JP Janeti for HJK Group
 # Dpt of Chemical Engineering, MIT
 
 #########################################################
@@ -26,9 +27,10 @@ def multitcgen(args,strfiles):
     else:
         jobdirs.append(tcgen(args,strfiles,method))
     # remove original files
-    for xyzf in strfiles:
-        os.remove(xyzf+'.xyz')
-        os.remove(xyzf+'.molinp')
+    if not args.jobdir:
+        for xyzf in strfiles:
+            os.remove(xyzf+'.xyz')
+            os.remove(xyzf+'.molinp')
     return jobdirs
 
 ### generate terachem input files ###
@@ -53,16 +55,17 @@ def tcgen(args,strfiles,method):
     # Overwrite plus add any new dictionary keys from commandline input.       
     for xyzf in strfiles:
         rdir = xyzf.rsplit('/',1)[0]
+        print("rdir is " + rdir)
         xyzft = xyzf.rsplit('/',1)[-1]
         xyzf += '.xyz'
         coordfs.append(xyzf.rsplit('/',1)[-1])
         coordname = xyzft
         # Setting jobname for files + truncated name for queue.
         if len(coordname) > 10:
-            nametrunc=coordname[0:6]+coordname[-4:]
+            nametrunc=coordname
         else:
             nametrunc=coordname
-        if not os.path.exists(rdir+'/'+nametrunc):
+        if not os.path.exists(rdir+'/'+nametrunc) and not args.jobdir:
             os.mkdir(rdir+'/'+nametrunc) 
         mdir = rdir+'/'+nametrunc
         if method:
@@ -73,9 +76,12 @@ def tcgen(args,strfiles,method):
             mdir = rdir+'/'+nametrunc+mmd
             if not os.path.exists(mdir):
                 os.mkdir(mdir)
-        jobdirs.append(mdir)
-        shutil.copy2(xyzf,mdir)
-        shutil.copy2(xyzf.replace('.xyz','.molinp'),mdir.replace('.xyz','.molinp'))
+        if not args.jobdir:
+            jobdirs.append(mdir)
+            shutil.copy2(xyzf,mdir)
+            shutil.copy2(xyzf.replace('.xyz','.molinp'),mdir.replace('.xyz','.molinp'))
+        elif args.jobdir:
+             jobdirs.append(rdir)
     # parse extra arguments
     # Method parsing, does not check if a garbage method is used here:
     unrestricted=False
@@ -141,14 +147,56 @@ def tcgen(args,strfiles,method):
        if not jobparams.has_key('levelshiftvalb'):
           jobparams['levelshiftvalb']='0.1'
     # Now we're ready to start building the input file
-    for i,jobd in enumerate(jobdirs):
-        output=open(jobd+'/terachem_input','w')
-        output.write('# file created with %s\n' % globs.PROGRAM)
-        jobparams['coordinates'] = coordfs[i]
-        for keys in jobparams.keys():
-            output.write('%s %s\n' %(keys,jobparams[keys]))
-        output.write('end\n')
-        output.close()
+    if not args.jobdir:
+        for i,jobd in enumerate(jobdirs):
+            output=open(jobd+'/terachem_input','w')
+            output.write('# file created with %s\n' % globs.PROGRAM)
+            jobparams['coordinates'] = coordfs[i]
+            for keys in jobparams.keys():
+                output.write('%s %s\n' %(keys,jobparams[keys]))
+            if  jobparams['run'] == 'minimize':
+                output.write('new_minimizer yes\n')
+                output.write('min_coordinates cartesian\n')
+            if args.tc_fix_dihedral:
+                temp = mol3D()
+                temp.readfromxyz(strfiles[i])
+                metal_ind = temp.findMetal()
+                fixed_atoms = list()
+                fixed_atoms = temp.getBondedAtoms(metal_ind)
+                fixed_atoms = [str(int(i)+1) for i in fixed_atoms] # 1-based indices
+                string_to_write = 'dihedral 0 ' + '_'.join(fixed_atoms)
+                print(string_to_write)
+                output.write('$constraint_set \n')
+                output.write(string_to_write + '\n')
+            output.write('end\n')
+            output.close()
+    elif args.jobdir:
+        for i,jobd in enumerate(jobdirs):
+            print('jobd is ' + jobd)
+            if args.name:
+                output=open(jobd+ '/'+args.name + '.in','w')
+            else:
+                output=open(jobd+'/terachem_input','w')
+            output.write('# file created with %s\n' % globs.PROGRAM)
+            jobparams['coordinates'] = coordfs[i]
+            for keys in jobparams.keys():
+                output.write('%s %s\n' %(keys,jobparams[keys]))
+            if  jobparams['run'] == 'minimize':
+                output.write('new_minimizer yes\n')
+                output.write('min_coordinates cartesian\n')
+            if args.tc_fix_dihedral:
+                temp = mol3D()
+                temp.readfromxyz(strfiles[i])
+                metal_ind = temp.findMetal()
+                fixed_atoms = list()
+                fixed_atoms = temp.getBondedAtoms(metal_ind)
+                fixed_atoms = [str(int(i)+1) for i in fixed_atoms] # 1-based indices
+                string_to_write = 'dihedral 0 ' + '_'.join(fixed_atoms)
+                print(string_to_write)
+                output.write('$constraint_set \n')
+                output.write(string_to_write + '\n')
+            output.write('end\n')
+            output.close()
     return jobdirs
 
 #####################################################
@@ -387,7 +435,7 @@ def qgen(args,strfiles,method):
             nametrunc=coordname[0:6]+coordname[-4:]
         else:
             nametrunc=coordname
-        if not os.path.exists(rdir+'/'+nametrunc):
+        if not os.path.exists(rdir+'/'+nametrunc) and not args.jobdir:
             os.mkdir(rdir+'/'+nametrunc) 
         mdir = rdir+'/'+nametrunc
         if method:
@@ -395,6 +443,7 @@ def qgen(args,strfiles,method):
             mdir = rdir+'/'+nametrunc+mmd
             if not os.path.exists(mdir):
                 os.mkdir(mdir)
+
         jobdirs.append(mdir)
         shutil.copy2(xyzf,mdir)
         shutil.copy2(xyzf.replace('.xyz','.molinp'),mdir.replace('.xyz','.molinp'))
@@ -451,5 +500,85 @@ def qgen(args,strfiles,method):
         # write $molecule block
         output.write('$molecule\n'+jobparams['CHARGE']+' '+jobparams['SPIN']+'\n')
         output.write(''.join(s)+'$end')
+        output.close()
+    return jobdirs
+
+
+
+    
+####################################################
+########## This module generates input  ############
+##########    for MOLPAC calculations   #############
+####################################################
+
+def mlpgen(args,strfiles,rootdir):
+    # get global variables
+    globs = globalvars()
+    jobdirs = []
+    coordfs = []
+    # Initialize the jobparams dictionary with mandatory/useful keywords.
+    jobparams = ['EF','PM7','XYZ','HESSIAN']
+    spin_keywords={1: 'SINGLET',
+                   2: 'DOUBLET',
+                   3: 'TRIPLET',
+                   4: 'QUARTET',
+                   5: 'QUINTET',
+                   6: 'SEXTET',
+                   7: 'SEPTET'}
+    # Overwrite plus add any new dictionary keys from commandline input.       
+    for xyzf in strfiles:
+        rdir = xyzf.rsplit('/',1)[0]
+        xyzft = xyzf.rsplit('/',1)[-1]
+        xyzf += '.xyz'
+        coordfs.append(xyzf.rsplit('/',1)[-1])
+        coordname = xyzft
+        # Setting jobname for files + truncated name for queue.
+        nametrunc=coordname
+        if not os.path.exists(rdir+'/'+nametrunc) and not args.jobdir:
+            os.mkdir(rdir+'/'+nametrunc) 
+        if args.jobdir:
+                mdir = args.jobdir
+        else:
+            mdir = rdir+'/'+nametrunc
+        jobdirs.append(mdir)
+#        shutil.copy2(xyzf,mdir)
+#        shutil.copy2(xyzf.replace('.xyz','.molinp'),mdir.replace('.xyz','.molinp'))
+    # Just carry over spin and charge keywords if they're set. Could do checks, none for now.
+    if args.spin:
+       jobparams.append(spin_keywords[int(args.spin)])
+       jobparams.append('UHF')
+
+    else:
+        jobparams.append("SINGLET")
+    if args.charge:
+       jobparams.append('CHARGE='+str(args.charge))
+    # Now we're ready to start building the input file and the job script
+    for i,jobd in enumerate(jobdirs):
+        output=open(strfiles[i] + '.mop','w')
+        f=open(strfiles[i]+'.xyz')
+        s = f.readlines()[2:] # read coordinates
+        f.close()
+        # write rem block
+        for terms in jobparams:
+            output.write(' '+ terms)
+        # write additional options
+        if (args.remoption):
+            if len(args.remoption)%2 > 0:
+                print 'WARNING: wrong number of arguments in -remoption'
+            else:
+                for elem in range(0,int(0.5*len(args.remoption))):
+                    key,val=args.remoption[2*elem],args.remoption[2*elem+1]
+                    output.write(key+'\t\t'+val+'\n')
+        output.write('\n' + nametrunc+'\n')
+        output.write('\n')
+        # write $molecule block
+        for lines in s:
+            ll = lines.split('\t')
+            for i,items in enumerate(ll):
+                output.write(' '+ items.strip('\n'))
+                if i>0:
+                    output.write(' 1')
+                if i == 3:
+                    output.write('\n')
         output.close()
     return jobdirs
